TailRecursion

Description

CMPS 112 : Programming Assignment 2

Do not change the skeleton code!

You may only replace the `error "TBD:..."` parts.

For this assignment, you may use the following library functions:

append (++)

Documentation

assoc :: Int -> String -> [(String, Int)] -> Int

`assoc def key [(k1,v1), (k2,v2), (k3,v3);...])`

searches the list for the first i such that ki = key. If such a ki is found, then vi is returned. Otherwise, if no such ki exists in the list, def is returned.

* your function should be tail recursive **

removeDuplicates :: [Int] -> [Int]

`removeDuplicates l`

returns the list of elements of l with duplicates that is, second, third ... occurrences, removed, and where the remaining elements appear in the same order as in l.

* your helper should be tail recursive **

for this problem only, you may use the library functions:

elem

wwhile :: (a -> (Bool, a)) -> a -> a

`wwhile f x` returns x' where there exist values

v_0,...,v_n such that

x is equal to v_0

x' is equal to v_n

for each i between `0` and `n-2`, we have `f v_i` equals `(true, v_i+1)`

`f v_n-1` equals `(false, v_n)`.

* your function should be tail recursive **

fixpointL :: (Int -> Int) -> Int -> [Int]

The **fixpoint** of a function f starting at x

`fixpoint f x` returns the FIRST element of the sequence x0, x1, x2, ...

x0 = x x1 = f x0 x2 = f x1 x3 = f x2 . . .

such that xn = f x_{n-1}

That is,

`fixpoint f x` should compute `f x` and then

IF x == f x then the fixpoint is x

OTHERWISE, the it is the (recursively computed) fixpoint of `f x`.

Fill in the implementation of `fixpointL f x0` which returns

the list [x_0, x_1, x_2, x_3, ... , x_n]

where

x = x_0

f x_0 = x_1, f x_1 = x_2, f x_2 = x_3, ... f x_n = x_{n+1}

xn = x_{n+1}

g :: Int -> Int

collatz :: Int -> Int

fixpointW :: (Int -> Int) -> Int -> Int

Now refactor your implementation of fixpointL so that it just returns the LAST element of the list, i.e. the xn that is equal to `f xn`